+++ ProjetLunaWeb/src/main/java/com/formation/dao/UserDao.java
@@ -10,9 +10,9 @@ public interface UserDao {
 
 	public void updateUser(User user);
 
-	public void removeUser(int userID);
+	public void removeUser(long userID);
 
-	public User getUser(int userID);
+	public User getUser(long userID);
 
 	public boolean getUserConnection(String user, String passwd);
