+++ ProjetLunaWeb/src/main/java/com/formation/service/ArticleService.java
@@ -9,9 +9,9 @@ public interface ArticleService {
 
 	public void updateArticle(Article article);
 
-	public void removeArticle(int articleID);
+	public void removeArticle(long articleID);
 
-	public Article getArticle(int articleID);
+	public Article getArticle(long articleID);
 
 	public List<Article> getAllArticle();
