+++ ProjetLunaWeb/src/main/java/com/formation/service/CommandeCreerService.java
@@ -8,8 +8,8 @@ import com.formation.persistence.CommandeCreer;
 public interface CommandeCreerService {
 	public void insertCommandeCreer(CommandeCreer commandeCreer);
 	public void updateCommandeCreer(CommandeCreer commandeCreer);
-	public void removeCommandeCreer(int commandeCreerID);
-	public CommandeCreer getCommandeCreer(int commandeCreerID);
+	public void removeCommandeCreer(long commandeCreerID);
+	public CommandeCreer getCommandeCreer(long commandeCreerID);
 	public List<CommandeCreer> getAllCommandeCreer();
 
 }
