+++ ProjetLunaWeb/src/main/java/com/formation/service/CommandeService.java
@@ -9,8 +9,8 @@ import com.formation.persistence.Commande;
 public interface CommandeService {
 	public void insertCommande(Commande commande);
 	public void updateCommande(Commande commande);
-	public void removeCommande(int commandeId);
-	public Commande getCommande(int commandeId);
+	public void removeCommande(long commandeId);
+	public Commande getCommande(long commandeId);
 	public List<Commande> getAllCommande();
 
 }
